requirements.txt
boto3
flask
twilio
gunicorn
python-dotenv